Broke out templates into their own file, broke out template generation & raw search, added generate query tests.
